Add 003 solution

+++ 001/001.cpp
@@ -4,6 +4,7 @@
 using namespace std;
 
 
+// returns ap summation of n
 unsigned long int s(unsigned long int n){
 	return n*(n+1)/2;
 }
@@ -13,7 +14,7 @@ int main(){
 	int t;
 	cin>>t;
 	for(int a0=0;a0<t;a0++){
-		int n;
+		unsigned long int n;
 		cin>>n;
 		n--;
 		cout<<3*s(n/3)+5*s(n/5)-15*s(n/15)<<endl;
